re-implemented friction models
