+++ src/friction_models/FrictionModels.jl
@@ -49,26 +49,6 @@ function get_friction_matrix end
 NQCModels.ndofs(model::ElectronicFrictionProvider) = model.ndofs
 NQCModels.dofs(model::ElectronicFrictionProvider) = 1:model.ndofs
 
-"""
-    ConstantFriction
-
-Friction model which returns a constant value for all positions. Use with a single value, a vector of diagonal values or a full-size Matrix
-
-"""
-struct ConstantFriction{T} <: ElectronicFrictionProvider
-    ndofs::Int
-    γ::T
-end
-
-"""
-    RandomFriction <: ElectronicFrictionProvider
-
-Provide a random positive semi-definite matrix of friction values.
-Used mostly for testing and examples.
-"""
-struct RandomFriction <: ElectronicFrictionProvider
-    ndofs::Int
-end
 
 """
     friction_matrix_indices(model, indices)
@@ -84,26 +64,6 @@ function get_friction_matrix end
 NQCModels.ndofs(model::ElectronicFrictionProvider) = model.ndofs
 NQCModels.dofs(model::ElectronicFrictionProvider) = 1:model.ndofs
 
-"""
-    ConstantFriction
-
-Friction model which returns a constant value for all positions. Use with a single value, a vector of diagonal values or a full-size Matrix
-
-"""
-struct ConstantFriction{T} <: ElectronicFrictionProvider
-    ndofs::Int
-    γ::T
-end
-
-"""
-    RandomFriction <: ElectronicFrictionProvider
-
-Provide a random positive semi-definite matrix of friction values.
-Used mostly for testing and examples.
-"""
-struct RandomFriction <: ElectronicFrictionProvider
-    ndofs::Int
-end
 
 """
     friction!(model::Model, F, R:AbstractMatrix)
@@ -155,9 +115,9 @@ export CompositeFrictionModel
 include("ase_friction_interface.jl")
 export ASEFrictionProvider
 
-# include("constant_friction.jl")
+include("constant_friction.jl")
 export ConstantFriction
-# include("random_friction.jl")
+include("random_friction.jl")
 export RandomFriction
 
 end # module
