+++ src/friction_models/random_friction.jl
@@ -1,6 +1,14 @@
 using Random: randn!
 
+"""
+    RandomFriction <: ElectronicFrictionProvider
 
+Provide a random positive semi-definite matrix of friction values.
+Used mostly for testing and examples.
+"""
+struct RandomFriction <: ElectronicFrictionProvider
+    ndofs::Int
+end
 
 function friction!(::RandomFriction, F::AbstractMatrix, ::AbstractMatrix)
     randn!(F)
@@ -8,3 +16,4 @@ function friction!(::RandomFriction, F::AbstractMatrix, ::AbstractMatrix)
     F .= (F + F')/2
 end
 
+NQCModels.ndofs(model::RandomFriction) = model.ndofs
